lirc_imon: add mouse and touchscreen input device support
Add an input device for the iMON PAD remote in mouse mode, and
on supported devices, a touchscreen input device. Also based
heavily on Rene Harder's iMON Touch patch submitted to the lirc
mailing list back in Nov. '08, but adapted to benefit all the
recent iMON devices.

Nb: does NOT yet work with the older ffdc devices, but it
shouldn't be terribly difficult to add support once I find
a guinea pig...