lirc_imon: add support for setting/enabling iMON internal clock via sysfs
lcdproc can be used to set the clock on an iMON device, but it can now
also be done by simply directing 'date' output into a sysfs attr.

To show a clock on your iMON device do the following:
    # date "+%y %m %d %w %H %M %S" > /sys/.../imon_clock

Requires that the display is of course present on your device, and that
it isn't currently open (by something like lcdproc's LCDd).

Tested successfully on a 15c2:ffdc VFD and 15c2:0045 LCD.

Patch primarily by Vitaly V. Bursov, I just added the bits to make it
work for the LCD devices too.