lirc_mceusb: explain the init process better
I've been sitting on a post from a Kevin Timmerman to
gmane.comp.hardware.lirc from back in early 2007 that details
a bunch of the mce device command strings and their meanings.
Transcribe that into comments in the code and give the various
strings more descriptive names. For example, 'pin_init1' is
actually 'get carrier mode and frequency'.

More detailed debugging spew that also utilizes this data soon
to follow...