lirc_mceusb: enable transmit on topseed 0x0008 devices
Special thanks to Greg Oliver for providing the hardware to
debug and figure out the proper intialization sequence for
this device.

Essentially, there seems to be a 3rd-gen class of mceusb devices,
which require different initialization from the 2nd-gen (and 1st-gen),
so I've created a new device table for those.

There's a quirk with a note about a usb reset being required for the
pinnacle device... That reset actually wedges the topseed 0x0008,
and the required magic for it is actually to send the mce device
reset command to the device prior to starting the init sequence. I
*suspect* this would also work in place of the crap being done for
the pinnacle device right now, but I don't have one to test that theory
out, so I'm leaving it be for now...

Nb: in the future, the multitude of tables here and in the imon driver
should be converted to device quirk tables, a la the upstream firewire
driver... (first example that comes to mind).