lirc_mceusb: override bInterval on topseed 0x0008 to 1
Complete and total testing involves actually yanking the device out
of the system, not just unloading and reloading the driver... The
topseed 0x0008 device comes up like so:

usb 2-9: config 1 interface 0 altsetting 0 endpoint 0x1 has an invalid bInterval 0, changing to 32
usb 2-9: config 1 interface 0 altsetting 0 endpoint 0x82 has an invalid bInterval 0, changing to 32

If we load the old lirc_mceusb, that got set to 1. Unload and load up an
updated lirc_mceusb to enable transmit, and it sticks at 1. Unplug and
replug with the updated driver and its back at 32, and the thing gets
completely wedged.

Upon inspection, the topseed 0x0001 and original ms mce transceiver also
come up with a stock bInterval of 0, so we're back to setting it to 1 for
everything but the pinnacle transceiver, where whatever its default is
seems to be required for functionality. Meh.