lircd: fix things up so decode behaves correctly
There were a few things I changed to __u32 earlier that should have been
left as unsigned long, and the old #ifdef __GLIBC__ hack appears to be
no longer necessary with ir_code as a __u64 (could be glibc got fixed,
not sure, but it works, so lets drop the cruft). Also converts a few
things that are type lirc_t that store return values from functions that
returned unsigned longs (now __u32's, mostly) over to __u32 types.

With this, lirc encode and decode behave 100% identical with 0.8.7
again, at least in testing on a 64-bit system...

Signed-off-by: Jarod Wilson <jarod@redhat.com>